refactor(terminal): extract terminal preview snapshot to separate module
Move terminal preview snapshot generation logic from terminalSurfaceRuntime.ts
to a new dedicated module terminalSurfacePreview.ts for better code organization
and maintainability.
